pyproject.toml
[project]
name = "thirukural-wallpaper-winos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.0.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
